Move save_to_json to F0 outputs (calles from C0 and E0)
